refactor(schema-utils): split ts definition generator
