+++ packages/schema-utils/src/definition-generator/index.ts
@@ -1,1 +1,4 @@
 export * from './TsDefinitionGenerator'
+export * from './AclDefinitionCodeGenerator'
+export * from './DefinitionCodeGenerator'
+export * from './DefinitionNamingConventions'
